使い回したい関数はutilに移動
