+++ src/js/RecentAddedExamples.vue
@@ -28,7 +28,7 @@
 <script>
 import axios from 'axios';
 import Example from './Example.vue';
-import NewExample from './util';
+import { NewExample } from './util';
 
 export default {
   data () {
